feat: break more sync functions
