+++ src/NextAudio.Matroska/MatroskaDemuxer.cs
@@ -73,74 +73,6 @@ public sealed partial class MatroskaDemuxer : AudioDemuxer
         return _currentBlockIndex < block.FrameCount;
     }
 
-    private void CheckCurrentStatePostSeek()
-    {
-        if (_position < _segmentElement.Position)
-        {
-            _segmentElement = default;
-        }
-
-
-        if (_currentClusterElement.HasValue)
-        {
-            if (_position < _currentClusterElement.Value.Position || _position > _currentClusterElement.Value.EndPosition)
-            {
-                _currentClusterElement = null;
-            }
-        }
-
-        if (_currentBlockGroupElement.HasValue)
-        {
-            if (_position < _currentBlockGroupElement.Value.Position || _position > _currentBlockGroupElement.Value.EndPosition)
-            {
-                _currentBlockGroupElement = null;
-            }
-        }
-
-        if (_currentBlock.HasValue)
-        {
-            if (_position < _currentBlock.Value.Element.Position || _position > _currentBlock.Value.Element.EndPosition)
-            {
-                _currentBlock = null;
-                _currentBlockIndex = 0;
-            }
-        }
-
-        if (!_currentBlock.HasValue)
-        {
-            return;
-        }
-
-        // If block was survived we will change the index relative to position
-        // and seek to the start of the frame.
-
-        var totalFrameSize = 0;
-        var blockStartPosition = _currentBlock.Value.Element.Position + _currentBlock.Value.Element.HeaderSize;
-
-        for (var i = 0; i < _currentBlock.Value.FrameCount; i++)
-        {
-            var frameSize = _currentBlock.Value.GetFrameSizeByIndex(i);
-
-            var frameStartPosition = totalFrameSize + blockStartPosition;
-
-            totalFrameSize += frameSize;
-
-            var frameEndPosition = totalFrameSize + blockStartPosition;
-
-            if (frameEndPosition < _position)
-            {
-                continue;
-            }
-
-            _currentBlockIndex = i;
-
-            if (frameStartPosition != _position)
-            {
-                _ = InternalSeek(frameStartPosition, SeekOrigin.Begin);
-            }
-        }
-    }
-
     /// <inheritdoc/>
     public override ValueTask<int> DemuxAsync(Memory<byte> buffer, CancellationToken cancellationToken = default)
     {
